+++ app/Http/Resources/GalleryConfigs/ZipChunkData.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) 2017-2018 Tobias Reich
+ * Copyright (c) 2018-2026 LycheeOrg.
+ */
+
+namespace App\Http\Resources\GalleryConfigs;
+
+use Spatie\LaravelData\Data;
+use Spatie\TypeScriptTransformer\Attributes\TypeScript;
+
+#[TypeScript()]
+class ZipChunkData extends Data
+{
+	public function __construct(
+		public int $total_chunks,
+		public int $total_photos,
+	) {
+	}
+}
