+++ resources/js/services/photo-service.ts
@@ -77,7 +77,7 @@ const PhotoService = {
 		photo_ids: string[],
 		from_id: string | null,
 		variant: App.Enum.DownloadVariantType,
-	): Promise<AxiosResponse<App.Http.Resources.GalleryConfigs.ZipChunksData>> {
+	): Promise<AxiosResponse<App.Http.Resources.GalleryConfigs.ZipChunkData>> {
 		return axios.get(`${Constants.getApiUrl()}Zip/chunks`, { params: { photo_ids: photo_ids.join(","), from_id: from_id, variant } });
 	},
